Add PCbi patch for PCem v17
This is a direct port of the patch for PCem v16. It has been tested to
compile for PCem and to run with PCbi, but not further.

Any modifications that are needed for PCem v17 specifically will be made
and committed later.

One known bug (from v16) is that PCbi's folder sharing doesn't refresh
automatically in PCem; you need to re-mount the ISO file. That issue may
or may not still exist with PCem v17.